src/main/java/de/jensknipper/greenmailexample/control/mail/MailSendingHandler.java
package de.jensknipper.greenmailexample.control.mail;

import org.springframework.beans.factory.annotation.Value;
import org.springframework.mail.SimpleMailMessage;
import org.springframework.mail.javamail.JavaMailSenderImpl;
import org.springframework.stereotype.Component;

@Component
public final class MailSendingHandler {

    @Value("${spring.mail.username}")
    private String mailUser;

    private final JavaMailSenderImpl javaMailSender;

    public MailSendingHandler(JavaMailSenderImpl javaMailSender) {
        this.javaMailSender = javaMailSender;
    }

    public void send(String recipient, String subject, String text) {
        SimpleMailMessage message = new SimpleMailMessage();
        message.setFrom(mailUser);
        message.setTo(recipient);
        message.setSubject(subject);
        message.setText(text);
        javaMailSender.send(message);
    }
}
